requirements.txt
torch==1.10.0
torchaudio==0.10.0
torchvision
torchcrepe
tensorboard
textgrid
Pillow
numpy
tqdm
jiwer
librosa
pandas
protobuf==3.20.0
git+https://github.com/ludlows/python-pesq#egg=pesq
psutil
pystoi
pytest
pyworld
pyyaml
matplotlib
seaborn
ipython
scipy
scikit-learn
ipywebrtc
argbind
sounddevice
keyboard